Rename SDK tool utilities
Use the tool-command-utils name consistently for the implementation and focused test suite.


Copilot-Session: a6568012-15e7-4b93-87b1-3d7aea7dffd7
